geminidocker/python-image/requirements.txt
openai
python-dotenv
gradio
google-generativeai